pyproject.toml
[build-system]
requires = ["hatchling>=1.5.0"]
build-backend = "hatchling.build"

[project]
name = "jupyter_ai_tools_exploration"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Framework :: Jupyter :: JupyterLab :: 4",
    "Framework :: Jupyter :: JupyterLab :: Extensions",
    "Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "jupyter_server>=2.4.0,<3",
    "jupyter_server_ai_tools",
    "jupyter_ai>=3.0.0a1",
    "langchain-aws",
    "langgraph",
]
version = "0.0.1"

[[project.authors]]
name = "Piyush Jain"
email = "piyushjain@duck.com"

[project.urls]
Homepage = "https://github.com/3coins/jupyter-ai-tools-exploration"

[project.optional-dependencies]
test = [
    "coverage",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-jupyter[server]>=0.6.0"
]

[tool.hatch.version]
path = "jupyter_ai_tools_exploration/_version.py"

[tool.hatch.build.targets.wheel.shared-data]
"jupyter-config/server-config" = "etc/jupyter/jupyter_server_config.d"

[tool.hatch.build.hooks.version]
path = "jupyter_ai_tools_exploration/_version.py"

[tool.jupyter-releaser.options]
version_cmd = "hatch version"

[tool.check-wheel-contents]
ignore = ["W002"]

[project.entry-points."jupyter_ai.personas"]
finance_persona = "jupyter_ai_tools_exploration.persona:TestPersona"
